move functions to hook and search into its own component
